requirements.txt
h2o==3.38.0.1
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.2
Pillow==9.3.0
psutil==5.8.0
scikit_learn==1.2.0
seaborn==0.11.2
shap==0.41.0
statsmodels==0.13.2
streamlit==1.15.1
